[Fix] Store TTT dh/dhb in fp32 to avoid fp16 overflow
The backward eta gradient at chunk-final positions can exceed the fp16
max (65504), turning into inf on store and NaN on the subsequent
inf×0 in chunk_bwd_kernel_dqke. bf16 has the range but not the
precision (8 mantissa bits vs fp16's 11), causing de ratio > 3%.
fp32 covers both range and precision; the kernel already accumulates
in fp32 and casts on store.

+++ fla/ops/ttt/chunk.py
@@ -989,8 +989,8 @@ def chunk_ttt_linear_bwd_norm(
     assert NK == 1, 'NK > 1 is not supported by TTT.'
     assert NV == 1, 'NV > 1 is not supported by TTT.'
 
-    dh = q.new_empty(B, NT, H, K, V)
-    dhb = q.new_empty(B, NT, H, 1, V)
+    dh = q.new_empty(B, NT, H, K, V, dtype=torch.float32)
+    dhb = q.new_empty(B, NT, H, 1, V, dtype=torch.float32)
     dh0 = torch.empty_like(h0, dtype=torch.float32) if h0 is not None else None
     dhb0 = torch.empty_like(hb0, dtype=torch.float32) if hb0 is not None else None
     dv = torch.empty_like(v)
